implement untextured rectangles
