+++ ps1-core/src/gpu/rasterizer/wgpuhardware.rs
@@ -1,1144 +1,34 @@
-use crate::gpu::gp0::{DrawSettings, SemiTransparencyMode, TextureColorDepthBits};
+mod blit;
+mod draw;
+
+use crate::gpu::gp0::DrawSettings;
+use crate::gpu::rasterizer::wgpuhardware::blit::{
+    CpuVramBlitPipeline, NativeScaledSyncPipeline, VramCopyPipeline, VramFillPipeline,
+};
+use crate::gpu::rasterizer::wgpuhardware::draw::DrawPipelines;
 use crate::gpu::rasterizer::{
     ClearPipeline, CpuVramBlitArgs, DrawLineArgs, DrawRectangleArgs, DrawTriangleArgs, FrameSize,
-    RasterizerInterface, TextureMappingMode, TriangleShading, TriangleTextureMapping,
-    VramVramBlitArgs,
+    RasterizerInterface, VramVramBlitArgs,
 };
 use crate::gpu::registers::Registers;
-use crate::gpu::{rasterizer, Color, Vertex, Vram, WgpuResources};
-use bytemuck::{Pod, Zeroable};
+use crate::gpu::{rasterizer, Color, Vram, WgpuResources};
 use std::collections::HashMap;
-use std::mem;
 use std::ops::Range;
 use std::rc::Rc;
-use wgpu::util::{BufferInitDescriptor, DeviceExt};
 use wgpu::{
-    BindGroup, BindGroupDescriptor, BindGroupEntry, BindGroupLayout, BindGroupLayoutDescriptor,
-    BindGroupLayoutEntry, BindingResource, BindingType, BlendComponent, BlendFactor,
-    BlendOperation, BlendState, Buffer, BufferBinding, BufferBindingType, BufferDescriptor,
-    BufferUsages, ColorTargetState, ColorWrites, CommandBuffer, CommandEncoder,
-    CommandEncoderDescriptor, ComputePass, ComputePassDescriptor, ComputePipeline,
-    ComputePipelineDescriptor, Device, Extent3d, FragmentState, FrontFace, ImageCopyTexture,
-    LoadOp, MultisampleState, Operations, Origin3d, PipelineCompilationOptions, PipelineLayout,
-    PipelineLayoutDescriptor, PolygonMode, PrimitiveState, PrimitiveTopology, PushConstantRange,
-    Queue, RenderPass, RenderPassColorAttachment, RenderPassDescriptor, RenderPipeline,
-    RenderPipelineDescriptor, ShaderModule, ShaderStages, StorageTextureAccess, StoreOp, Texture,
-    TextureAspect, TextureDescriptor, TextureDimension, TextureFormat, TextureUsages,
-    TextureViewDescriptor, TextureViewDimension, VertexAttribute, VertexBufferLayout, VertexState,
-    VertexStepMode,
+    BindGroup, Buffer, CommandBuffer, CommandEncoder, CommandEncoderDescriptor,
+    ComputePassDescriptor, Device, Extent3d, ImageCopyTexture, LoadOp, Operations, Origin3d, Queue,
+    RenderPassColorAttachment, RenderPassDescriptor, StoreOp, Texture, TextureAspect,
+    TextureDescriptor, TextureDimension, TextureFormat, TextureUsages, TextureViewDescriptor,
 };
 
 const VRAM_WIDTH: u32 = 1024;
 const VRAM_HEIGHT: u32 = 512;
 
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct ShaderDrawSettings {
-    draw_area_top_left: [f32; 2],
-    draw_area_bottom_right: [f32; 2],
-    force_mask_bit: u32,
-}
-
-impl ShaderDrawSettings {
-    fn new(draw_settings: &DrawSettings, resolution_scale: u32) -> Self {
-        let resolution_scale = resolution_scale as i32;
-
-        Self {
-            draw_area_top_left: [
-                (resolution_scale * draw_settings.draw_area_top_left.x) as f32,
-                (resolution_scale * draw_settings.draw_area_top_left.y) as f32,
-            ],
-            draw_area_bottom_right: [
-                (resolution_scale * (draw_settings.draw_area_bottom_right.x + 1)) as f32,
-                (resolution_scale * (draw_settings.draw_area_bottom_right.y + 1)) as f32,
-            ],
-            force_mask_bit: draw_settings.force_mask_bit.into(),
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct UntexturedVertex {
-    position: [i32; 2],
-    color: [u32; 3],
-}
-
-impl UntexturedVertex {
-    const ATTRIBUTES: [VertexAttribute; 2] = wgpu::vertex_attr_array![0 => Sint32x2, 1 => Uint32x3];
-
-    const LAYOUT: VertexBufferLayout<'static> = VertexBufferLayout {
-        array_stride: mem::size_of::<Self>() as u64,
-        step_mode: VertexStepMode::Vertex,
-        attributes: &Self::ATTRIBUTES,
-    };
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct TexturedVertex {
-    position: [i32; 2],
-    color: [u32; 3],
-    uv: [u32; 2],
-    texpage: [u32; 2],
-    tex_window_mask: [u32; 2],
-    tex_window_offset: [u32; 2],
-    clut: [u32; 2],
-    color_depth: u32,
-    modulated: u32,
-}
-
-impl TexturedVertex {
-    const ATTRIBUTES: [VertexAttribute; 9] = wgpu::vertex_attr_array![
-        0 => Sint32x2,
-        1 => Uint32x3,
-        2 => Uint32x2,
-        3 => Uint32x2,
-        4 => Uint32x2,
-        5 => Uint32x2,
-        6 => Uint32x2,
-        7 => Uint32,
-        8 => Uint32,
-    ];
-
-    const LAYOUT: VertexBufferLayout<'static> = VertexBufferLayout {
-        array_stride: mem::size_of::<Self>() as u64,
-        step_mode: VertexStepMode::Vertex,
-        attributes: &Self::ATTRIBUTES,
-    };
-
-    fn new(
-        position: [i32; 2],
-        color: Color,
-        u: u8,
-        v: u8,
-        texture_mapping: &TriangleTextureMapping,
-    ) -> Self {
-        Self {
-            position,
-            color: [color.r.into(), color.g.into(), color.b.into()],
-            uv: [u.into(), v.into()],
-            texpage: [64 * texture_mapping.texpage.x_base, texture_mapping.texpage.y_base],
-            tex_window_mask: [
-                texture_mapping.window.x_mask << 3,
-                texture_mapping.window.y_mask << 3,
-            ],
-            tex_window_offset: [
-                texture_mapping.window.x_offset << 3,
-                texture_mapping.window.y_offset << 3,
-            ],
-            clut: [(16 * texture_mapping.clut_x).into(), texture_mapping.clut_y.into()],
-            color_depth: match texture_mapping.texpage.color_depth {
-                TextureColorDepthBits::Four => 0,
-                TextureColorDepthBits::Eight => 1,
-                TextureColorDepthBits::Fifteen => 2,
-            },
-            modulated: (texture_mapping.mode == TextureMappingMode::Modulated).into(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum TrianglePipeline {
-    Untextured(Option<SemiTransparencyMode>),
-    Textured(Option<SemiTransparencyMode>),
-}
-
-#[derive(Debug)]
-struct DrawBatch {
-    draw_settings: DrawSettings,
-    pipeline: TrianglePipeline,
-    start: u32,
-    end: u32,
-}
-
-#[derive(Debug)]
-struct TrianglePipelines {
-    ram_untextured_buffer: Vec<UntexturedVertex>,
-    untextured_buffer: Buffer,
-    untextured_opaque_pipeline: RenderPipeline,
-    untextured_average_pipeline: RenderPipeline,
-    untextured_add_pipeline: RenderPipeline,
-    untextured_subtract_pipeline: RenderPipeline,
-    untextured_add_quarter_pipeline: RenderPipeline,
-    ram_textured_buffer: Vec<TexturedVertex>,
-    textured_buffer: Buffer,
-    textured_bind_group: BindGroup,
-    textured_opaque_pipeline: RenderPipeline,
-    textured_average_pipeline: RenderPipeline,
-    textured_add_pipeline: RenderPipeline,
-    textured_subtract_pipeline_opaque: RenderPipeline,
-    textured_subtract_pipeline_transparent: RenderPipeline,
-    textured_add_quarter_pipeline: RenderPipeline,
-    batches: Vec<DrawBatch>,
-}
-
-fn create_untextured_triangle_pipeline(
-    device: &Device,
-    draw_shader: &ShaderModule,
-    fs_entry_point: &str,
-    pipeline_layout: &PipelineLayout,
-    blend: BlendState,
-) -> RenderPipeline {
-    device.create_render_pipeline(&RenderPipelineDescriptor {
-        label: "untextured_opaque_triangle_pipeline".into(),
-        layout: Some(&pipeline_layout),
-        vertex: VertexState {
-            module: draw_shader,
-            entry_point: "vs_untextured",
-            compilation_options: PipelineCompilationOptions::default(),
-            buffers: &[UntexturedVertex::LAYOUT],
-        },
-        primitive: PrimitiveState {
-            topology: PrimitiveTopology::TriangleList,
-            strip_index_format: None,
-            front_face: FrontFace::Ccw,
-            cull_mode: None,
-            unclipped_depth: false,
-            polygon_mode: PolygonMode::Fill,
-            conservative: false,
-        },
-        depth_stencil: None,
-        multisample: MultisampleState::default(),
-        fragment: Some(FragmentState {
-            module: draw_shader,
-            entry_point: fs_entry_point,
-            compilation_options: PipelineCompilationOptions::default(),
-            targets: &[Some(ColorTargetState {
-                format: TextureFormat::Rgba8Unorm,
-                blend: Some(blend),
-                write_mask: ColorWrites::ALL,
-            })],
-        }),
-        multiview: None,
-    })
-}
-
-fn create_textured_triangle_pipeline(
-    device: &Device,
-    draw_shader: &ShaderModule,
-    fs_entry_point: &str,
-    pipeline_layout: &PipelineLayout,
-    blend: BlendState,
-) -> RenderPipeline {
-    device.create_render_pipeline(&RenderPipelineDescriptor {
-        label: "textured_opaque_triangle_pipeline".into(),
-        layout: Some(&pipeline_layout),
-        vertex: VertexState {
-            module: draw_shader,
-            entry_point: "vs_textured",
-            compilation_options: PipelineCompilationOptions::default(),
-            buffers: &[TexturedVertex::LAYOUT],
-        },
-        primitive: PrimitiveState {
-            topology: PrimitiveTopology::TriangleList,
-            strip_index_format: None,
-            front_face: FrontFace::Ccw,
-            cull_mode: None,
-            unclipped_depth: false,
-            polygon_mode: PolygonMode::Fill,
-            conservative: false,
-        },
-        depth_stencil: None,
-        multisample: MultisampleState::default(),
-        fragment: Some(FragmentState {
-            module: draw_shader,
-            entry_point: fs_entry_point,
-            compilation_options: PipelineCompilationOptions::default(),
-            targets: &[Some(ColorTargetState {
-                format: TextureFormat::Rgba8Unorm,
-                blend: Some(blend),
-                write_mask: ColorWrites::ALL,
-            })],
-        }),
-        multiview: None,
-    })
-}
-
-impl TrianglePipelines {
-    const MAX_VERTICES: u64 = 15000;
-
-    fn new(device: &Device, draw_shader: &ShaderModule, native_vram: &Texture) -> Self {
-        let untextured_buffer = device.create_buffer(&BufferDescriptor {
-            label: "untextured_opaque_triangle_vertex_buffer".into(),
-            size: Self::MAX_VERTICES * mem::size_of::<UntexturedVertex>() as u64,
-            usage: BufferUsages::COPY_DST | BufferUsages::VERTEX,
-            mapped_at_creation: false,
-        });
-
-        let untextured_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "untextured_opaque_triangle_pipeline_layout".into(),
-            bind_group_layouts: &[],
-            push_constant_ranges: &[PushConstantRange {
-                stages: ShaderStages::FRAGMENT,
-                range: 0..mem::size_of::<ShaderDrawSettings>() as u32,
-            }],
-        });
-
-        let untextured_opaque_pipeline = create_untextured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_untextured_opaque",
-            &untextured_layout,
-            BlendState::REPLACE,
-        );
-
-        let untextured_average_pipeline = create_untextured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_untextured_average",
-            &untextured_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::Src1Alpha,
-                    dst_factor: BlendFactor::OneMinusSrc1Alpha,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let untextured_add_pipeline = create_untextured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_untextured_opaque",
-            &untextured_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::One,
-                    dst_factor: BlendFactor::One,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let untextured_subtract_pipeline = create_untextured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_untextured_opaque",
-            &untextured_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::One,
-                    dst_factor: BlendFactor::One,
-                    operation: BlendOperation::ReverseSubtract,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let untextured_add_quarter_pipeline = create_untextured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_untextured_add_quarter",
-            &untextured_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::Src1Alpha,
-                    dst_factor: BlendFactor::One,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let textured_buffer = device.create_buffer(&BufferDescriptor {
-            label: "textured_opaque_triangle_vertex_buffer".into(),
-            size: Self::MAX_VERTICES * mem::size_of::<TexturedVertex>() as u64,
-            usage: BufferUsages::COPY_DST | BufferUsages::VERTEX,
-            mapped_at_creation: false,
-        });
-
-        let textured_bind_group_layout =
-            device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-                label: "textured_opaque_triangle_bind_group_layout".into(),
-                entries: &[BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::StorageTexture {
-                        access: StorageTextureAccess::ReadOnly,
-                        format: native_vram.format(),
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                }],
-            });
-
-        let native_vram_view = native_vram.create_view(&TextureViewDescriptor::default());
-        let textured_bind_group = device.create_bind_group(&BindGroupDescriptor {
-            label: "textured_opaque_triangle_bind_group".into(),
-            layout: &textured_bind_group_layout,
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: BindingResource::TextureView(&native_vram_view),
-            }],
-        });
-
-        let textured_pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "textured_opaque_triangle_pipeline_layout".into(),
-            bind_group_layouts: &[&textured_bind_group_layout],
-            push_constant_ranges: &[PushConstantRange {
-                stages: ShaderStages::FRAGMENT,
-                range: 0..mem::size_of::<ShaderDrawSettings>() as u32,
-            }],
-        });
-
-        let textured_opaque_pipeline = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_opaque",
-            &textured_pipeline_layout,
-            BlendState::REPLACE,
-        );
-
-        let textured_average_pipeline = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_average",
-            &textured_pipeline_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::Src1Alpha,
-                    dst_factor: BlendFactor::OneMinusSrc1Alpha,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let textured_add_pipeline = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_add",
-            &textured_pipeline_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::One,
-                    dst_factor: BlendFactor::Src1Alpha,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let textured_subtract_pipeline_opaque = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_subtract_opaque_texels",
-            &textured_pipeline_layout,
-            BlendState::REPLACE,
-        );
-
-        let textured_subtract_pipeline_transparent = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_subtract_transparent_texels",
-            &textured_pipeline_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::One,
-                    dst_factor: BlendFactor::One,
-                    operation: BlendOperation::ReverseSubtract,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        let textured_add_quarter_pipeline = create_textured_triangle_pipeline(
-            device,
-            draw_shader,
-            "fs_textured_add_quarter",
-            &textured_pipeline_layout,
-            BlendState {
-                color: BlendComponent {
-                    src_factor: BlendFactor::One,
-                    dst_factor: BlendFactor::Src1Alpha,
-                    operation: BlendOperation::Add,
-                },
-                alpha: BlendComponent::REPLACE,
-            },
-        );
-
-        Self {
-            ram_untextured_buffer: Vec::with_capacity(Self::MAX_VERTICES as usize),
-            untextured_buffer,
-            untextured_opaque_pipeline,
-            untextured_average_pipeline,
-            untextured_add_pipeline,
-            untextured_subtract_pipeline,
-            untextured_add_quarter_pipeline,
-            ram_textured_buffer: Vec::with_capacity(Self::MAX_VERTICES as usize),
-            textured_buffer,
-            textured_bind_group,
-            textured_opaque_pipeline,
-            textured_average_pipeline,
-            textured_add_pipeline,
-            textured_subtract_pipeline_opaque,
-            textured_subtract_pipeline_transparent,
-            textured_add_quarter_pipeline,
-            batches: Vec::with_capacity(Self::MAX_VERTICES as usize),
-        }
-    }
-
-    fn add_untextured_triangle(
-        &mut self,
-        vertices: [Vertex; 3],
-        shading: TriangleShading,
-        semi_transparency_mode: Option<SemiTransparencyMode>,
-        draw_settings: &DrawSettings,
-    ) {
-        let pipeline = TrianglePipeline::Untextured(semi_transparency_mode);
-
-        if !self.batches.last().is_some_and(|batch| {
-            &batch.draw_settings == draw_settings && batch.pipeline == pipeline
-        }) {
-            let start = self.ram_untextured_buffer.len() as u32;
-            self.batches.push(DrawBatch {
-                draw_settings: draw_settings.clone(),
-                pipeline,
-                start,
-                end: start,
-            });
-        }
-
-        let draw_offset = draw_settings.draw_offset;
-        let positions = vertices.map(|vertex| [vertex.x + draw_offset.x, vertex.y + draw_offset.y]);
-
-        let colors = match shading {
-            TriangleShading::Flat(color) => [color; 3],
-            TriangleShading::Gouraud(colors) => colors,
-        };
-
-        for (position, color) in positions.into_iter().zip(colors) {
-            self.ram_untextured_buffer.push(UntexturedVertex {
-                position,
-                color: [color.r.into(), color.g.into(), color.b.into()],
-            });
-        }
-
-        self.batches.last_mut().unwrap().end += 3;
-    }
-
-    fn add_textured_triangle(
-        &mut self,
-        vertices: [Vertex; 3],
-        shading: TriangleShading,
-        semi_transparency_mode: Option<SemiTransparencyMode>,
-        texture_mapping: &TriangleTextureMapping,
-        draw_settings: &DrawSettings,
-    ) {
-        let pipeline = TrianglePipeline::Textured(semi_transparency_mode);
-
-        if semi_transparency_mode == Some(SemiTransparencyMode::Subtract)
-            || !self.batches.last().is_some_and(|batch| {
-                &batch.draw_settings == draw_settings && batch.pipeline == pipeline
-            })
-        {
-            let start = self.ram_textured_buffer.len() as u32;
-            self.batches.push(DrawBatch {
-                draw_settings: draw_settings.clone(),
-                pipeline,
-                start,
-                end: start,
-            });
-        }
-
-        let draw_offset = draw_settings.draw_offset;
-        let positions = vertices.map(|vertex| [vertex.x + draw_offset.x, vertex.y + draw_offset.y]);
-
-        let colors = match shading {
-            TriangleShading::Flat(color) => [color; 3],
-            TriangleShading::Gouraud(colors) => colors,
-        };
-
-        for (i, position) in positions.into_iter().enumerate() {
-            self.ram_textured_buffer.push(TexturedVertex::new(
-                position,
-                colors[i],
-                texture_mapping.u[i],
-                texture_mapping.v[i],
-                texture_mapping,
-            ));
-        }
-
-        self.batches.last_mut().unwrap().end += 3;
-    }
-
-    fn prepare(&mut self, queue: &Queue) {
-        log::debug!(
-            "Preparing to draw {} untextured opaque triangle vertices",
-            self.ram_untextured_buffer.len()
-        );
-        log::debug!(
-            "Preparing to draw {} textured opaque triangle vertices",
-            self.ram_textured_buffer.len()
-        );
-
-        queue.write_buffer(
-            &self.untextured_buffer,
-            0,
-            bytemuck::cast_slice(&self.ram_untextured_buffer),
-        );
-        queue.write_buffer(
-            &self.textured_buffer,
-            0,
-            bytemuck::cast_slice(&self.ram_textured_buffer),
-        );
-
-        self.ram_untextured_buffer.clear();
-        self.ram_textured_buffer.clear();
-    }
-
-    fn draw<'rpass>(&'rpass mut self, resolution_scale: u32, render_pass: &mut RenderPass<'rpass>) {
-        log::debug!("Executing {} untextured opaque triangle batches", self.batches.len());
-
-        for batch in self.batches.drain(..) {
-            let draw_settings = ShaderDrawSettings::new(&batch.draw_settings, resolution_scale);
-
-            match batch.pipeline {
-                TrianglePipeline::Untextured(semi_transparency_mode) => {
-                    let pipeline = match semi_transparency_mode {
-                        Some(SemiTransparencyMode::Average) => &self.untextured_average_pipeline,
-                        Some(SemiTransparencyMode::Add) => &self.untextured_add_pipeline,
-                        Some(SemiTransparencyMode::Subtract) => &self.untextured_subtract_pipeline,
-                        Some(SemiTransparencyMode::AddQuarter) => {
-                            &self.untextured_add_quarter_pipeline
-                        }
-                        None => &self.untextured_opaque_pipeline,
-                    };
-
-                    render_pass.set_pipeline(pipeline);
-                    render_pass.set_push_constants(
-                        ShaderStages::FRAGMENT,
-                        0,
-                        bytemuck::cast_slice(&[draw_settings]),
-                    );
-                    render_pass.set_vertex_buffer(0, self.untextured_buffer.slice(..));
-
-                    render_pass.draw(batch.start..batch.end, 0..1);
-                }
-                TrianglePipeline::Textured(Some(SemiTransparencyMode::Subtract)) => {
-                    render_pass.set_pipeline(&self.textured_subtract_pipeline_opaque);
-                    render_pass.set_push_constants(
-                        ShaderStages::FRAGMENT,
-                        0,
-                        bytemuck::cast_slice(&[draw_settings]),
-                    );
-                    render_pass.set_bind_group(0, &self.textured_bind_group, &[]);
-                    render_pass.set_vertex_buffer(0, self.textured_buffer.slice(..));
-
-                    render_pass.draw(batch.start..batch.end, 0..1);
-
-                    render_pass.set_pipeline(&self.textured_subtract_pipeline_transparent);
-                    render_pass.set_push_constants(
-                        ShaderStages::FRAGMENT,
-                        0,
-                        bytemuck::cast_slice(&[draw_settings]),
-                    );
-
-                    render_pass.draw(batch.start..batch.end, 0..1);
-                }
-                TrianglePipeline::Textured(semi_transparency_mode) => {
-                    let pipeline = match semi_transparency_mode {
-                        Some(SemiTransparencyMode::Average) => &self.textured_average_pipeline,
-                        Some(SemiTransparencyMode::Add) => &self.textured_add_pipeline,
-                        Some(SemiTransparencyMode::Subtract) => unreachable!(),
-                        Some(SemiTransparencyMode::AddQuarter) => {
-                            &self.textured_add_quarter_pipeline
-                        }
-                        None => &self.textured_opaque_pipeline,
-                    };
-
-                    render_pass.set_pipeline(pipeline);
-                    render_pass.set_push_constants(
-                        ShaderStages::FRAGMENT,
-                        0,
-                        bytemuck::cast_slice(&[draw_settings]),
-                    );
-                    render_pass.set_bind_group(0, &self.textured_bind_group, &[]);
-                    render_pass.set_vertex_buffer(0, self.textured_buffer.slice(..));
-
-                    render_pass.draw(batch.start..batch.end, 0..1);
-                }
-            }
-        }
-    }
-}
-
-// Must match CpuVramBlitArgs in cpuvramblit.wgsl
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct ShaderCpuVramBlitArgs {
-    position: [u32; 2],
-    size: [u32; 2],
-    force_mask_bit: u32,
-    check_mask_bit: u32,
-}
-
-#[derive(Debug)]
-struct CpuVramBlitPipeline {
-    ram_buffer: Vec<u32>,
-    bind_group_0: BindGroup,
-    bind_group_layout_1: BindGroupLayout,
-    pipeline: ComputePipeline,
-}
-
-impl CpuVramBlitPipeline {
-    // Must match X/Y workgroup size in shader
-    const WORKGROUP_SIZE: u32 = 16;
-
-    fn new(device: &Device, native_vram: &Texture) -> Self {
-        let bind_group_layout_0 = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: "cpu_vram_blit_bind_group_layout_0".into(),
-            entries: &[BindGroupLayoutEntry {
-                binding: 0,
-                visibility: ShaderStages::COMPUTE,
-                ty: BindingType::StorageTexture {
-                    access: StorageTextureAccess::ReadWrite,
-                    format: native_vram.format(),
-                    view_dimension: TextureViewDimension::D2,
-                },
-                count: None,
-            }],
-        });
-
-        let bind_group_layout_1 = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: "cpu_vram_blit_bind_group_layout_1".into(),
-            entries: &[BindGroupLayoutEntry {
-                binding: 0,
-                visibility: ShaderStages::COMPUTE,
-                ty: BindingType::Buffer {
-                    ty: BufferBindingType::Storage { read_only: true },
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            }],
-        });
-
-        let native_vram_view = native_vram.create_view(&TextureViewDescriptor::default());
-        let bind_group_0 = device.create_bind_group(&BindGroupDescriptor {
-            label: "cpu_vram_blit_bind_group".into(),
-            layout: &bind_group_layout_0,
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: BindingResource::TextureView(&native_vram_view),
-            }],
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "cpu_vram_blit_pipeline_layout".into(),
-            bind_group_layouts: &[&bind_group_layout_0, &bind_group_layout_1],
-            push_constant_ranges: &[PushConstantRange {
-                stages: ShaderStages::COMPUTE,
-                range: 0..mem::size_of::<ShaderCpuVramBlitArgs>() as u32,
-            }],
-        });
-
-        let shader =
-            device.create_shader_module(wgpu::include_wgsl!("wgpuhardware/cpuvramblit.wgsl"));
-        let pipeline = device.create_compute_pipeline(&ComputePipelineDescriptor {
-            label: "cpu_vram_blit_pipeline".into(),
-            layout: Some(&pipeline_layout),
-            module: &shader,
-            entry_point: "cpu_vram_blit",
-            compilation_options: PipelineCompilationOptions::default(),
-        });
-
-        Self {
-            ram_buffer: Vec::with_capacity((VRAM_WIDTH * VRAM_HEIGHT) as usize),
-            bind_group_0,
-            bind_group_layout_1,
-            pipeline,
-        }
-    }
-
-    fn prepare(&mut self, device: &Device, args: &CpuVramBlitArgs, buffer: &[u16]) -> BindGroup {
-        let copy_len = (args.width * args.height) as usize;
-
-        self.ram_buffer.clear();
-        self.ram_buffer.extend(buffer.iter().copied().map(u32::from).take(copy_len));
-
-        let buffer = device.create_buffer_init(&BufferInitDescriptor {
-            label: "cpu_vram_blit_buffer".into(),
-            contents: bytemuck::cast_slice(&self.ram_buffer),
-            usage: BufferUsages::STORAGE,
-        });
-
-        device.create_bind_group(&BindGroupDescriptor {
-            label: "cpu_vram_blit_bind_group_1".into(),
-            layout: &self.bind_group_layout_1,
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: BindingResource::Buffer(BufferBinding {
-                    buffer: &buffer,
-                    offset: 0,
-                    size: None,
-                }),
-            }],
-        })
-    }
-
-    fn dispatch<'cpass>(
-        &'cpass self,
-        args: &CpuVramBlitArgs,
-        bind_group_1: &'cpass BindGroup,
-        compute_pass: &mut ComputePass<'cpass>,
-    ) {
-        let shader_args = ShaderCpuVramBlitArgs {
-            position: [args.x, args.y],
-            size: [args.width, args.height],
-            force_mask_bit: args.force_mask_bit.into(),
-            check_mask_bit: args.check_mask_bit.into(),
-        };
-
-        compute_pass.set_pipeline(&self.pipeline);
-        compute_pass.set_bind_group(0, &self.bind_group_0, &[]);
-        compute_pass.set_bind_group(1, &bind_group_1, &[]);
-        compute_pass.set_push_constants(0, bytemuck::cast_slice(&[shader_args]));
-
-        let x_groups =
-            args.width / Self::WORKGROUP_SIZE + u32::from(args.width % Self::WORKGROUP_SIZE != 0);
-        let y_groups =
-            args.height / Self::WORKGROUP_SIZE + u32::from(args.height % Self::WORKGROUP_SIZE != 0);
-        compute_pass.dispatch_workgroups(x_groups, y_groups, 1);
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct ShaderVramCopyArgs {
-    source: [u32; 2],
-    destination: [u32; 2],
-    size: [u32; 2],
-    force_mask_bit: u32,
-    check_mask_bit: u32,
-}
-
-impl ShaderVramCopyArgs {
-    fn new(args: &VramVramBlitArgs) -> Self {
-        Self {
-            source: [args.source_x, args.source_y],
-            destination: [args.dest_x, args.dest_y],
-            size: [args.width, args.height],
-            force_mask_bit: args.force_mask_bit.into(),
-            check_mask_bit: args.check_mask_bit.into(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct VramCopyPipeline {
-    bind_group: BindGroup,
-    pipeline: ComputePipeline,
-}
-
-impl VramCopyPipeline {
-    const X_WORKGROUP_SIZE: u32 = 16;
-
-    fn new(device: &Device, native_vram: &Texture) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: "vram_copy_bind_group_layout".into(),
-            entries: &[BindGroupLayoutEntry {
-                binding: 0,
-                visibility: ShaderStages::COMPUTE,
-                ty: BindingType::StorageTexture {
-                    access: StorageTextureAccess::ReadWrite,
-                    format: native_vram.format(),
-                    view_dimension: TextureViewDimension::D2,
-                },
-                count: None,
-            }],
-        });
-
-        let native_vram_view = native_vram.create_view(&TextureViewDescriptor::default());
-        let bind_group = device.create_bind_group(&BindGroupDescriptor {
-            label: "vram_copy_bind_group".into(),
-            layout: &bind_group_layout,
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: BindingResource::TextureView(&native_vram_view),
-            }],
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "vram_copy_pipeline_layout".into(),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[PushConstantRange {
-                stages: ShaderStages::COMPUTE,
-                range: 0..mem::size_of::<ShaderVramCopyArgs>() as u32,
-            }],
-        });
-
-        let shader = device.create_shader_module(wgpu::include_wgsl!("wgpuhardware/vramcopy.wgsl"));
-        let pipeline = device.create_compute_pipeline(&ComputePipelineDescriptor {
-            label: "vram_copy_pipeline".into(),
-            layout: Some(&pipeline_layout),
-            module: &shader,
-            entry_point: "vram_copy",
-            compilation_options: PipelineCompilationOptions::default(),
-        });
-
-        Self { bind_group, pipeline }
-    }
-
-    fn dispatch<'cpass>(
-        &'cpass self,
-        args: &VramVramBlitArgs,
-        compute_pass: &mut ComputePass<'cpass>,
-    ) {
-        let vram_copy_args = ShaderVramCopyArgs::new(args);
-
-        compute_pass.set_pipeline(&self.pipeline);
-        compute_pass.set_push_constants(0, bytemuck::cast_slice(&[vram_copy_args]));
-        compute_pass.set_bind_group(0, &self.bind_group, &[]);
-
-        let x_workgroups = args.width / Self::X_WORKGROUP_SIZE
-            + u32::from(args.width % Self::X_WORKGROUP_SIZE != 0);
-        compute_pass.dispatch_workgroups(x_workgroups, args.height, 1);
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct ShaderVramFillArgs {
-    position: [u32; 2],
-    size: [u32; 2],
-    color: u32,
-}
-
-#[derive(Debug)]
-struct VramFillPipeline {
-    bind_group: BindGroup,
-    pipeline: ComputePipeline,
-}
-
-impl VramFillPipeline {
-    const WORKGROUP_SIZE: u32 = 16;
-
-    fn new(device: &Device, native_vram: &Texture) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: "vram_fill_bind_group_layout".into(),
-            entries: &[BindGroupLayoutEntry {
-                binding: 0,
-                visibility: ShaderStages::COMPUTE,
-                ty: BindingType::StorageTexture {
-                    access: StorageTextureAccess::WriteOnly,
-                    format: native_vram.format(),
-                    view_dimension: TextureViewDimension::D2,
-                },
-                count: None,
-            }],
-        });
-
-        let native_vram_view = native_vram.create_view(&TextureViewDescriptor::default());
-        let bind_group = device.create_bind_group(&BindGroupDescriptor {
-            label: "vram_fill_bind_group".into(),
-            layout: &bind_group_layout,
-            entries: &[BindGroupEntry {
-                binding: 0,
-                resource: BindingResource::TextureView(&native_vram_view),
-            }],
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "vram_fill_pipeline_layout".into(),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[PushConstantRange {
-                stages: ShaderStages::COMPUTE,
-                range: 0..mem::size_of::<ShaderVramFillArgs>() as u32,
-            }],
-        });
-
-        let shader = device.create_shader_module(wgpu::include_wgsl!("wgpuhardware/vramfill.wgsl"));
-        let pipeline = device.create_compute_pipeline(&ComputePipelineDescriptor {
-            label: "vram_fill_pipeline".into(),
-            layout: Some(&pipeline_layout),
-            module: &shader,
-            entry_point: "vram_fill",
-            compilation_options: PipelineCompilationOptions::default(),
-        });
-
-        Self { bind_group, pipeline }
-    }
-
-    fn dispatch<'cpass>(
-        &'cpass self,
-        x: u32,
-        y: u32,
-        width: u32,
-        height: u32,
-        color: Color,
-        compute_pass: &mut ComputePass<'cpass>,
-    ) {
-        let args = ShaderVramFillArgs {
-            position: [x, y],
-            size: [width, height],
-            color: u32::from(color.r >> 3)
-                | (u32::from(color.g >> 3) << 5)
-                | (u32::from(color.b >> 3) << 10),
-        };
-
-        compute_pass.set_pipeline(&self.pipeline);
-        compute_pass.set_bind_group(0, &self.bind_group, &[]);
-        compute_pass.set_push_constants(0, bytemuck::cast_slice(&[args]));
-
-        let x_workgroups =
-            width / Self::WORKGROUP_SIZE + u32::from(width % Self::WORKGROUP_SIZE != 0);
-        let y_workgroups =
-            height / Self::WORKGROUP_SIZE + u32::from(height % Self::WORKGROUP_SIZE != 0);
-        compute_pass.dispatch_workgroups(x_workgroups, y_workgroups, 1);
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Zeroable, Pod)]
-struct VramSyncVertex {
-    position: [i32; 2],
-}
-
-impl VramSyncVertex {
-    const ATTRIBUTES: [VertexAttribute; 1] = wgpu::vertex_attr_array![0 => Sint32x2];
-
-    const LAYOUT: VertexBufferLayout<'static> = VertexBufferLayout {
-        array_stride: mem::size_of::<Self>() as u64,
-        step_mode: VertexStepMode::Vertex,
-        attributes: &Self::ATTRIBUTES,
-    };
-}
-
-#[derive(Debug)]
-struct NativeScaledSyncPipeline {
-    bind_group: BindGroup,
-    pipeline: RenderPipeline,
-}
-
-impl NativeScaledSyncPipeline {
-    fn new(device: &Device, native_vram: &Texture, resolution_scale: u32) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
-            label: "native_scaled_sync_bind_group_layout".into(),
-            entries: &[
-                BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::StorageTexture {
-                        access: StorageTextureAccess::ReadOnly,
-                        format: native_vram.format(),
-                        view_dimension: TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-                BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: ShaderStages::FRAGMENT,
-                    ty: BindingType::Buffer {
-                        ty: BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                },
-            ],
-        });
-
-        let native_vram_view = native_vram.create_view(&TextureViewDescriptor::default());
-        let resolution_scale_buffer = device.create_buffer_init(&BufferInitDescriptor {
-            label: "native_scaled_sync_buffer".into(),
-            contents: &resolution_scale.to_le_bytes(),
-            usage: BufferUsages::UNIFORM,
-        });
-        let bind_group = device.create_bind_group(&BindGroupDescriptor {
-            label: "native_scaled_sync_bind_group".into(),
-            layout: &bind_group_layout,
-            entries: &[
-                BindGroupEntry {
-                    binding: 0,
-                    resource: BindingResource::TextureView(&native_vram_view),
-                },
-                BindGroupEntry {
-                    binding: 1,
-                    resource: BindingResource::Buffer(BufferBinding {
-                        buffer: &resolution_scale_buffer,
-                        offset: 0,
-                        size: None,
-                    }),
-                },
-            ],
-        });
-
-        let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
-            label: "native_scaled_sync_pipeline_layout".into(),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[],
-        });
-
-        let shader = device.create_shader_module(wgpu::include_wgsl!("wgpuhardware/vramsync.wgsl"));
-        let pipeline = device.create_render_pipeline(&RenderPipelineDescriptor {
-            label: "native_scaled_sync_pipeline".into(),
-            layout: Some(&pipeline_layout),
-            vertex: VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                compilation_options: PipelineCompilationOptions::default(),
-                buffers: &[VramSyncVertex::LAYOUT],
-            },
-            primitive: PrimitiveState {
-                topology: PrimitiveTopology::TriangleStrip,
-                strip_index_format: None,
-                front_face: FrontFace::Ccw,
-                cull_mode: None,
-                unclipped_depth: false,
-                polygon_mode: PolygonMode::Fill,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: MultisampleState::default(),
-            fragment: Some(FragmentState {
-                module: &shader,
-                entry_point: "native_to_scaled",
-                compilation_options: PipelineCompilationOptions::default(),
-                targets: &[Some(ColorTargetState {
-                    format: TextureFormat::Rgba8Unorm,
-                    blend: Some(BlendState::REPLACE),
-                    write_mask: ColorWrites::ALL,
-                })],
-            }),
-            multiview: None,
-        });
-
-        Self { bind_group, pipeline }
-    }
-
-    fn prepare(&self, device: &Device, position: [u32; 2], size: [u32; 2]) -> Buffer {
-        let position = position.map(|n| n as i32);
-        let size = size.map(|n| n as i32);
-
-        let vertices = [
-            VramSyncVertex { position: [position[0], position[1]] },
-            VramSyncVertex { position: [position[0] + size[0], position[1]] },
-            VramSyncVertex { position: [position[0], position[1] + size[1]] },
-            VramSyncVertex { position: [position[0] + size[0], position[1] + size[1]] },
-        ];
-
-        device.create_buffer_init(&BufferInitDescriptor {
-            label: "native_scaled_sync_vertex_buffer".into(),
-            contents: bytemuck::cast_slice(&vertices),
-            usage: BufferUsages::VERTEX,
-        })
-    }
-
-    fn draw<'rpass>(&'rpass self, buffer: &'rpass Buffer, render_pass: &mut RenderPass<'rpass>) {
-        render_pass.set_pipeline(&self.pipeline);
-        render_pass.set_bind_group(0, &self.bind_group, &[]);
-        render_pass.set_vertex_buffer(0, buffer.slice(..));
-        render_pass.draw(0..4, 0..1);
-    }
-}
-
 #[derive(Debug)]
 enum DrawCommand {
     DrawTriangle { args: DrawTriangleArgs, draw_settings: DrawSettings },
+    DrawRectangle { args: DrawRectangleArgs, draw_settings: DrawSettings },
     CpuVramBlit { args: CpuVramBlitArgs, buffer_bind_group: BindGroup, sync_vertex_buffer: Buffer },
     VramCopy { args: VramVramBlitArgs, sync_vertex_buffer: Buffer },
     VramFill { x: u32, y: u32, width: u32, height: u32, color: Color, sync_vertex_buffer: Buffer },
@@ -1159,7 +49,7 @@ pub struct WgpuRasterizer {
     native_vram: Texture,
     frame_textures: HashMap<FrameSize, Texture>,
     clear_pipeline: ClearPipeline,
-    triangle_pipelines: TrianglePipelines,
+    draw_pipelines: DrawPipelines,
     cpu_vram_blit_pipeline: CpuVramBlitPipeline,
     vram_copy_pipeline: VramCopyPipeline,
     vram_fill_pipeline: VramFillPipeline,
@@ -1207,8 +97,7 @@ impl WgpuRasterizer {
 
         let draw_shader =
             device.create_shader_module(wgpu::include_wgsl!("wgpuhardware/draw.wgsl"));
-        let untextured_opaque_triangle_pipeline =
-            TrianglePipelines::new(&device, &draw_shader, &native_vram);
+        let draw_pipelines = DrawPipelines::new(&device, &draw_shader, &native_vram);
 
         let cpu_vram_blit_pipeline = CpuVramBlitPipeline::new(&device, &native_vram);
 
@@ -1227,7 +116,7 @@ impl WgpuRasterizer {
             native_vram,
             frame_textures: HashMap::with_capacity(20),
             clear_pipeline,
-            triangle_pipelines: untextured_opaque_triangle_pipeline,
+            draw_pipelines,
             cpu_vram_blit_pipeline,
             vram_copy_pipeline,
             vram_fill_pipeline,
@@ -1265,15 +154,18 @@ impl WgpuRasterizer {
         let mut i = 0;
         while let Some(command) = self.draw_commands.get(i) {
             match command {
-                DrawCommand::DrawTriangle { .. } => {
+                DrawCommand::DrawTriangle { .. } | DrawCommand::DrawRectangle { .. } => {
                     let mut j = i + 1;
                     while j < self.draw_commands.len()
-                        && matches!(&self.draw_commands[j], DrawCommand::DrawTriangle { .. })
+                        && matches!(
+                            &self.draw_commands[j],
+                            DrawCommand::DrawTriangle { .. } | DrawCommand::DrawRectangle { .. }
+                        )
                     {
                         j += 1;
                     }
 
-                    self.execute_draw_triangles(i..j, &mut encoder);
+                    self.execute_draw(i..j, &mut encoder);
 
                     i = j;
                 }
@@ -1299,41 +191,27 @@ impl WgpuRasterizer {
         Some(encoder.finish())
     }
 
-    fn execute_draw_triangles(
-        &mut self,
-        draw_command_range: Range<usize>,
-        encoder: &mut CommandEncoder,
-    ) {
+    fn execute_draw(&mut self, draw_command_range: Range<usize>, encoder: &mut CommandEncoder) {
         log::debug!(
             "Executing {} draw triangle commands",
             draw_command_range.end - draw_command_range.start
         );
 
         for draw_command in &self.draw_commands[draw_command_range.clone()] {
-            let DrawCommand::DrawTriangle { args, draw_settings } = draw_command else { continue };
-
-            match &args.texture_mapping {
-                Some(texture_mapping) => {
-                    self.triangle_pipelines.add_textured_triangle(
-                        args.vertices,
-                        args.shading,
-                        args.semi_transparent.then_some(args.semi_transparency_mode),
-                        texture_mapping,
-                        draw_settings,
-                    );
+            match draw_command {
+                DrawCommand::DrawTriangle { args, draw_settings } => {
+                    self.draw_pipelines.add_triangle(args, draw_settings);
                 }
-                None => {
-                    self.triangle_pipelines.add_untextured_triangle(
-                        args.vertices,
-                        args.shading,
-                        args.semi_transparent.then_some(args.semi_transparency_mode),
-                        draw_settings,
-                    );
+                DrawCommand::DrawRectangle { args, draw_settings } => {
+                    self.draw_pipelines.add_rectangle(args, draw_settings);
                 }
+                DrawCommand::CpuVramBlit { .. }
+                | DrawCommand::VramFill { .. }
+                | DrawCommand::VramCopy { .. } => {}
             }
         }
 
-        self.triangle_pipelines.prepare(&self.queue);
+        let draw_buffers = self.draw_pipelines.prepare(&self.device);
 
         let scaled_vram_view = self.scaled_vram.create_view(&TextureViewDescriptor::default());
         {
@@ -1347,7 +225,7 @@ impl WgpuRasterizer {
                 ..RenderPassDescriptor::default()
             });
 
-            self.triangle_pipelines.draw(self.resolution_scale, &mut render_pass);
+            self.draw_pipelines.draw(&draw_buffers, self.resolution_scale, &mut render_pass);
         }
     }
 
@@ -1384,7 +262,7 @@ impl WgpuRasterizer {
                             &mut compute_pass,
                         );
                     }
-                    DrawCommand::DrawTriangle { .. } => {}
+                    DrawCommand::DrawTriangle { .. } | DrawCommand::DrawRectangle { .. } => {}
                 }
             }
         }
@@ -1408,7 +286,7 @@ impl WgpuRasterizer {
                     | DrawCommand::VramFill { sync_vertex_buffer, .. } => {
                         self.native_scaled_sync_pipeline.draw(sync_vertex_buffer, &mut render_pass);
                     }
-                    DrawCommand::DrawTriangle { .. } => {}
+                    DrawCommand::DrawTriangle { .. } | DrawCommand::DrawRectangle { .. } => {}
                 }
             }
         }
@@ -1423,7 +301,10 @@ impl RasterizerInterface for WgpuRasterizer {
 
     fn draw_line(&mut self, _args: DrawLineArgs, _draw_settings: &DrawSettings) {}
 
-    fn draw_rectangle(&mut self, _args: DrawRectangleArgs, _draw_settings: &DrawSettings) {}
+    fn draw_rectangle(&mut self, args: DrawRectangleArgs, draw_settings: &DrawSettings) {
+        self.draw_commands
+            .push(DrawCommand::DrawRectangle { args, draw_settings: draw_settings.clone() });
+    }
 
     fn vram_fill(&mut self, x: u32, y: u32, width: u32, height: u32, color: Color) {
         // TODO scaled/native sync
